Enable CodeAnalysis and Stypecop for ProxyUtilities
