+++ cf-dotnet-sdk.nuspec
@@ -48,7 +48,7 @@
     <file src="lib\loggregator-net45\CloudFoundry.Loggregator.Client.XML"     target="lib\Net45\CloudFoundry.Loggregator.Client.XML" />
     <file src="lib\uaa-net45\CloudFoundry.UAA.Client.dll"                     target="lib\Net45\CloudFoundry.UAA.Client.dll" />
     <file src="lib\uaa-net45\CloudFoundry.UAA.Client.xml"                     target="lib\Net45\CloudFoundry.UAA.Client.XML" />
-    <file src="lib\proxyutils-net45\WebSocket4Net.ProxyUtils.dll"             target="lib\Net45\WebSocket4Net.ProxyUtils.dll" />
+    <file src="lib\proxyutils-net45\WebSocket4Net.ProxyUtilities.dll"         target="lib\Net45\WebSocket4Net.ProxyUtilities.dll" />
 
     <file src="lib\CloudFoundry.CloudController.V2.Client.dll"                target="lib\Win8\CloudFoundry.CloudController.V2.Client.dll" />
     <file src="lib\CloudFoundry.CloudController.V2.Client.xml"                target="lib\Win8\CloudFoundry.CloudController.V2.Client.xml" />
