+++ cf-dotnet-sdk.sln
@@ -48,7 +48,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CloudFoundry.Loggregator.Cl
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CloudFoundry.CloudController.Common.Net45", "src\CloudFoundry.CloudController.Common.Net45\CloudFoundry.CloudController.Common.Net45.csproj", "{126AEDA2-FA3B-42EB-B728-770A975DEF2B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebSocket4Net.ProxyUtils", "src\WebSocket4Net.ProxyUtils\WebSocket4Net.ProxyUtils.csproj", "{446818CB-1AA8-449A-B58A-01DA64A36871}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebSocket4Net.ProxyUtilities", "src\WebSocket4Net.ProxyUtilities\WebSocket4Net.ProxyUtilities.csproj", "{446818CB-1AA8-449A-B58A-01DA64A36871}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
