+++ src/CloudFoundry.Loggregator.Client.Net45/CloudFoundry.Loggregator.Client.Net45.csproj
@@ -127,9 +127,9 @@
       <Project>{126aeda2-fa3b-42eb-b728-770a975def2b}</Project>
       <Name>CloudFoundry.CloudController.Common.Net45</Name>
     </ProjectReference>
-    <ProjectReference Include="..\WebSocket4Net.ProxyUtils\WebSocket4Net.ProxyUtils.csproj">
+    <ProjectReference Include="..\WebSocket4Net.ProxyUtilities\WebSocket4Net.ProxyUtilities.csproj">
       <Project>{446818cb-1aa8-449a-b58a-01da64a36871}</Project>
-      <Name>WebSocket4Net.ProxyUtils</Name>
+      <Name>WebSocket4Net.ProxyUtilities</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
