+++ src/WebSocket4Net.ProxyUtilities/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="StyleCop.MSBuild" version="4.7.49.1" targetFramework="net45" developmentDependency="true" />
   <package id="WebSocket4Net" version="0.12" targetFramework="net45" />
 </packages>
